requirements.txt
numpy==1.24.3
pandas==1.5.3
pandas==2.0.3
Pillow==9.4.0
Pillow==10.0.1
pydrive==1.3.1
PyDrive2==1.17.0
PyPDF2==3.0.1
pytube==15.0.0
rembg==2.0.50
st_paywall==0.1.5
streamlit==1.27.0
streamlit_option_menu==0.3.6
